Add ability to use ThemeChooser only to pick a theme, without also applying it.
This modification gives the ability to use the ThemeChooser one of two different
ways.  It may be started to choose a new active theme that is applied to the
device immediately after it is chosen.  It may also be started only to pick a
theme whose URI will be returned to the caller without applying it to the device.

This patch also introduces a new class with static methods to build the
Intents and start the ThemeChooser appropriately based on the desired
outcome.

Signed-off-by: Josh Guilfoyle <josh.guilfoyle@t-mobile.com>